requirements.txt
Django>5.0,<5.2
# 파이썬은 3.13 고정
# 사용가능한버전 : 3.10, 3.11, 3.12, 3.13
djangorestframework==3.15.0,<3.15.2
psycopg2>=2.9.9,<2.10.0
drf-spectacular>=0.26.5,<0.27.0
django-extensions>=3.2.0,<3.3.0
django-cors-headers
djangorestframework-simplejwt>=5.0.0,<6.0.0